Service 1: refactor CLI code for re-use in service
